Lager tester for TilgangUnderPilotering og fikser sjekken på kontorer i pilot-liste
