+++ src/main/java/no/nav/tag/tiltaksgjennomforing/integrasjon/configurationProperties/PilotProperties.java
@@ -3,8 +3,6 @@ package no.nav.tag.tiltaksgjennomforing.integrasjon.configurationProperties;
 import lombok.Data;
 import no.nav.tag.tiltaksgjennomforing.domene.NavEnhet;
 import no.nav.tag.tiltaksgjennomforing.domene.NavIdent;
-import no.nav.tag.tiltaksgjennomforing.domene.exceptions.TilgangskontrollException;
-import no.nav.tag.tiltaksgjennomforing.integrasjon.axsys.AxsysService;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -15,18 +13,9 @@ import java.util.List;
 @Component
 @ConfigurationProperties(prefix = "tiltaksgjennomforing.pilot")
 public class PilotProperties {
+
     private boolean enabled;
-    private List<String> enheter = new ArrayList<>();
+    private List<NavEnhet> enheter = new ArrayList<>();
     private List<NavIdent> identer = new ArrayList<>();
-    private final AxsysService axsysService;
 
-    public void sjekkTilgang(NavIdent ident) {
-        if (enabled && !identer.contains(ident)) {
-            List<NavEnhet> enheterInnloggetVeilederHarTilgangTil = axsysService.hentEnheterVeilederHarTilgangTil(ident);
-            if (!enheter.contains(enheterInnloggetVeilederHarTilgangTil)) {
-                throw new TilgangskontrollException("Ident " + ident.asString() + " er ikke lagt til i lista over brukere med tilgang.");
-            }
-
-        }
-    }
 }
